# SparseList Follow Protocol Speedup